COMP: Fixed warnings.

+++ Servers/Filters/vtkTransferFunctionEditorWidgetSimple1D.cxx
@@ -1144,7 +1144,7 @@ void vtkTransferFunctionEditorWidgetSimple1D::SetElementHSVColor(
 void vtkTransferFunctionEditorWidgetSimple1D::SetElementScalar(
   unsigned int idx, double value)
 {
-  unsigned int size = this->Nodes->size();
+  unsigned int size = static_cast<unsigned int>(this->Nodes->size());
   if (idx >= this->Nodes->size())
     {
     return;
